Remove sequence.h into residue.h \ Update residue_exchange string encode DNA tpye check
