+++ src/include/dawg/matic.h
@@ -11,7 +11,7 @@
 #include <dawg/rate.h>
 #include <dawg/wood.h>
 #include <dawg/root.h>
-#include <dawg/sequence.h>
+#include <dawg/residue.h>
 #include <dawg/ma.h>
 
 #include <vector>
@@ -26,7 +26,7 @@ namespace details {
 struct indel_data {
 	typedef std::pair<double, boost::uint32_t> element;
 	typedef std::stack<element> stack;
-	
+
 	stack ins;
 	stack del;
 
@@ -45,9 +45,9 @@ public:
 	wood usertree;
 	wood_meta_type metatree;
 	double tree_scale;
-	
+
 	bool gap_overlap;
-	
+
 	subst_model     sub_mod;
 	rate_model      rat_mod;
 	root_model      rut_mod;
@@ -55,7 +55,7 @@ public:
 	indel_model del_mod;
 
 	residue::data_type gap_base;
-	
+
 	void evolve(sequence &child, indel_data &indels, double T, residue::data_type branch_color,
 		sequence::const_iterator first, sequence::const_iterator last,
 		mutt &m) const;
@@ -65,7 +65,7 @@ public:
 	evolve_indels(sequence &child, indel_data &indels, double T, residue::data_type branch_color,
 		sequence::const_iterator first, sequence::const_iterator last,
 		mutt &m) const;
-		
+
 	inline boost::uint32_t next_indel(double d, double &f, bool bDel) const {
 		double ins_rate = ins_mod.rate();
 		double del_rate = del_mod.rate();
@@ -85,7 +85,7 @@ public:
 		f -= del_rate;
 		return x + ((bDel) ? 4 : 2);
 	}
-	
+
 	inline boost::uint32_t mark_del(boost::uint32_t u, sequence &child,
 			sequence::const_iterator &first, sequence::const_iterator last) const {
 		boost::uint32_t uu;
@@ -131,7 +131,7 @@ public:
 	inline void clear_configuration() {
 		configs.clear();
 	}
-	
+
 	template<class It>
 	inline bool configure(It first, It last) {
 		clear_configuration();
@@ -144,11 +144,11 @@ public:
 
 	// Run the simulation
 	void walk(alignment& aln);
-	
+
 	// Precalculate stuff for simulation
 	void pre_walk(alignment& aln);
-	
-	
+
+
 	template<typename _It>
 	void seed(_It first, _It last) {
 		maxx.seed(first, last);
@@ -157,33 +157,33 @@ public:
 	void seed(T t) {
 		maxx.seed(t);
 	}
-	
+
 	matic() : branch_color(0)
 	{}
-	
+
 protected:
 	typedef dawg::details::matic_section section;
 	struct segment : public std::vector<std::unique_ptr<section>> {
 		residue_exchange rex;
 	};
 	typedef std::vector<segment> segment_vector;
-	
+
 	typedef std::map<std::string, wood::data_type::size_type> label_to_index_type;
 	typedef std::vector<details::sequence_data> seq_buffers_type;
-	
+
 	seq_buffers_type seqs;
-	
+
 	segment_vector configs;
 	mutt maxx;
-	
+
 	residue::data_type branch_color;
-	
+
 	label_to_index_type label_union;
 	alignment::size_type aln_size;
-	
+
 	bool add_config_section(const dawg::ma &ma);
 	bool finalize_configuration();
-	
+
 	void align(alignment& aln, const seq_buffers_type &seqs, const residue_exchange &rex);
 };
